[flang2] Cast _N_WHO expression to int; NFCI
The macro _N_WHO in main.cpp counts the number of elements in the static
array who, and is used in array bounds and loop bounds. Changing its
type from size_t to int silences a number of -Wsign-compare warnings.

+++ tools/flang2/flang2exe/main.cpp
@@ -88,7 +88,7 @@ static action_map_t *phase_dump_map;
  */
 static char *who[] = {"init", "import",   "expand", "", "",
                       "",     "assemble", "xref",   ""};
-#define _N_WHO (sizeof(who) / sizeof(char *))
+#define _N_WHO (int)(sizeof(who) / sizeof(char *))
 static INT xtimes[_N_WHO];
 static char *cmdline = NULL;
 static char *ccff_filename = NULL;
